+++ src/main/java/roomescape/domain/repository/ReservationRepository.java
@@ -1,4 +1,6 @@
-package roomescape.domain;
+package roomescape.domain.repository;
+
+import roomescape.domain.Reservation;
 
 import java.util.List;
